+++ frontend/src/components/form/api/ApiCheckbox.vue
@@ -1,20 +1,21 @@
 <!--
-Displays a field as a time picker + write access via API wrapper
+Displays a field as a v-checkbox + write access via API wrapper
 -->
 
 <template>
   <api-wrapper
     v-bind="$props">
     <template slot-scope="wrapper">
-      <time-picker
-        :value="wrapper.localValue"
+      <v-checkbox
+        :input-value="wrapper.localValue"
         v-bind="$attrs"
         :readonly="readonly"
         :disabled="disabled"
         :error-messages="wrapper.errorMessages"
-        @input="wrapper.on.input">
+        hide-details="auto"
+        @change="wrapper.on.input">
         <status-icon slot="append" :status="wrapper.status" />
-      </time-picker>
+      </v-checkbox>
     </template>
   </api-wrapper>
 </template>
@@ -22,20 +23,18 @@ Displays a field as a time picker + write access via API wrapper
 <script>
 import { apiPropsMixin } from '@/mixins/apiPropsMixin'
 import ApiWrapper from './ApiWrapper'
-import StatusIcon from './StatusIcon'
-import TimePicker from './TimePicker'
+import StatusIcon from '../base/StatusIcon'
 
 export default {
-  name: 'ApiTimePicker',
-  components: { ApiWrapper, StatusIcon, TimePicker },
+  name: 'ApiCheckbox',
+  components: { ApiWrapper, StatusIcon },
   mixins: [apiPropsMixin],
   props: {
     // disable delay per default
     autoSaveDelay: { type: Number, default: 0, required: false }
   },
   data () {
-    return {
-    }
+    return {}
   }
 }
 </script>
