+++ frontend/src/components/form/base/ColorPicker.vue
