+++ frontend/src/components/form/base/StatusIcon.vue
